se hizo validaciones en el delete y update password

+++ app/users/users-services-delete.php
@@ -6,16 +6,26 @@ class usersAPI
     function deleteUser($datos)
     {
 
-        $id = $datos['iduser'];
 
-        $objDB = new ExtraerDatos();
 
-        $ejecucion = $objDB->deleteUser($id);
+        if ($datos) {
 
-        if ($ejecucion) {
-            echo json_encode(array("data" => null, "error" => "0", "msg" => "Usuario eliminado :)",));
+            if (isset($datos['iduser'])) {
+                $id = $datos['iduser'];
+                $objDB = new ExtraerDatos();
+
+                $ejecucion = $objDB->deleteUser($id);
+
+                if ($ejecucion) {
+                    echo json_encode(array("data" => null, "error" => "0", "msg" => "Usuario eliminado :)",));
+                } else {
+                    echo json_encode(array("data" => null, "error" => "1", "msg" => "No se pudo eliminar :(",));
+                }
+            } else {
+                echo json_encode(array("data" => null, "error" => "1", "msg" => "Falta el iduser",));
+            }
         } else {
-            echo json_encode(array("data" => null, "error" => "1", "msg" => "No se pudo eliminar :(",));
+            echo json_encode(array("data" => null, "error" => "1", "msg" => "Faltan datos",));
         }
     }
 }
